refactor: className, confirm delete and struct

+++ public/static/js/admin/videos-index.js
@@ -9,20 +9,27 @@ const videosIndex = function () {
       const videoId = btnDeleteVideo.dataset.id;
       const url = btnDeleteVideo.dataset.url;
 
-      lotv.ajax({
-        method: 'DELETE',
-        url,
-        success: res => {
-          if (res?.success) {
-            lotv.toast.fire({
-              title: 'Success',
-              message: res.message,
-              type: 'success',
-            });
-            document.getElementById(`video${videoId}`).remove();
-          }
+      const dialog = confirmDialog({
+        title: 'Confirm Deletion',
+        message: 'Are you sure you want to delete this video? This action cannot be undone.',
+        onConfirm: () => {
+          lotv.ajax({
+            method: 'DELETE',
+            url,
+            success: res => {
+              if (res?.success) {
+                lotv.toast.fire({
+                  title: 'Success',
+                  message: res.message,
+                  type: 'success',
+                });
+                document.getElementById(`video${videoId}`).remove();
+              }
+            },
+          });
         },
       });
+      dialog.show();
     });
   }
 };
